Update how isLast is set when data is updated
Check if the current slide should be the last slide when dataStream is updated.
This fixes an issue where the dataStream is updated but the carousel
slides do not move. In the current state, isLast will incorrectly be set
to false. With these changes, isLast will stay false if the carousel is
    still on the last slide after the data updates